added dynamic packet slowing
Slows down the packets based on the input time. Made the time more accurate between given and output by reducing the fuzzing sleep times. The packets will now calculate the rate they should be sent out to meet the given time and added more verbose feedback. Use case would be to run in listener mode for a while to assess ARP traffic then edit the runtime based on the volume of traffic.